backend/requirements.txt
# FastAPI web framework
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0

# Database
SQLAlchemy>=2.0.25
alembic>=1.13.0
pymongo>=4.6.0 # includes bson

# Authentication and Security
PyJWT>=2.8.0
bcrypt>=4.1.0
python-multipart>=0.0.9

# Environment Configuration
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Development Tools
pytest>=7.4.0
httpx>=0.24.1
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0

# Email Validation
email-validator

# WebSocket
websockets==12.0